% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangularize.R
\name{rectangularize}
\alias{rectangularize}
\title{Bind JSON data parsed as an R list to a data frame and unnest nested columns}
\usage{
rectangularize(content)
}
\arguments{
\item{content}{A list containing JSON data parsed into R format.}
}
\value{
A data frame with JSON data bound and all nested list columns unnested
to achieve a tidy format.
}
\description{
This function takes JSON data parsed as an R list, binds it to a data frame, and
then recursively applies the "unnest_recursively" function to unnest any nested
list columns in the data frame. The input data can contain either a single observation
(a list representing a row) or multiple observations (a list of lists representing
multiple rows).
}
\examples{
# Sample JSON data parsed as an R list

}
